uwaPySense/requirements.txt
pyodbc
pyserial